[feat] 소식지 [fix] 커뮤니티 컴포넌트 분리
